Added ParallelBulkClient
ParalleBulkClient: A client side handler which asks jobs in bulk, processes them in parallel and pushes them back to the server in bulk.
BulkTest: A basic test app to see if ParallelBulkClient works fine. This uses the same jar files and jobs as the SimpleClient

Along with these, two custom exception classes and a few changes to the previous files such that the bulk client